third_party/WebKit/LayoutTests/fast/dom/StyleSheet/detached-parent-rule-without-wrapper.html

<html>
<head>
<style>
@media all { #test { color: green; } }
</style>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<div id="test"></div>
<script>
description("Test that detached CSSOM subtrees keeps a reference to CSS stylesheets lacking wrappers.");

shouldBe('getMatchedCSSRules(document.getElementById("test")).length', '1');
rule = getMatchedCSSRules(document.getElementById("test")).item(0);
debug("Removing &lt;style>...");
document.head.removeChild(document.getElementsByTagName("style")[0]);
shouldNotBe("rule.parentRule", "null");
shouldBe('getMatchedCSSRules(document.getElementById("test"))', 'null');
</script>
</body>
</html>
