third_party/WebKit/LayoutTests/fast/dom/StyleSheet/detached-parent-rule-without-wrapper-expected.txt

CONSOLE WARNING: line 1: document.getMatchedCSSRules() is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4606972603138048 for more details.
Test that detached CSSOM subtrees keeps a reference to CSS stylesheets lacking wrappers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getMatchedCSSRules(document.getElementById("test")).length is 1
Removing <style>...
PASS rule.parentRule is not null
PASS getMatchedCSSRules(document.getElementById("test")) is null
PASS successfullyParsed is true

TEST COMPLETE
